Expand an extension in place, as the wireframe has it
The sheet was wrong: the design expands a row into its own settings, and
that is what this does again. The rows an open extension adds are siblings
in the same section rather than a nested layout, so each is a real form
row — label column, control styling, hover and focus all the system's.

A text field in a form row collapses to nothing beside a long label, which
is how "Custom Brew Executable Path" ended up invisible but still
clickable. Fields carry a border and a fixed width now.

Commands are grouped together after the preferences instead of each
sitting in a section of its own, and adding is one menu in the section
header rather than three buttons trailing the list.
